[CoinMarketCap] Move adapters class to proper package
